backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pandas
cors 